refactor: split yields out into shareYields() report
Yields live on a different indexer entity (token_snapshot) than prices
(token_instance_snapshot), are populated by different events, and have
their own failure modes. Bundling them into sharePrices() meant a slow
or erroring yields query stalled the whole report — visible as the
dashboard chart hanging on its loader because combineLatest waited for
both queries before emitting.

Splits yields into its own pool.reports.shareYields() report that mirrors
sharePrices' shape ({ timestamp, shareClasses: Record<id, ShareYields> }),
so consumers that don't need yields don't pay for the extra query and
the chart can render prices independently of yields.
