+++ src/entities/Reports/PoolReports.ts
@@ -1,6 +1,7 @@
 import { Centrifuge } from '../../Centrifuge.js'
 import { Entity } from '../Entity.js'
 import { Pool } from '../Pool.js'
+import { PoolShareYieldsReport, ShareYieldsReportFilter } from './PoolShareYieldsReport.js'
 import { PoolSharePricesReport, SharePricesReportFilter } from './PoolSharePricesReport.js'
 
 export class PoolReports extends Entity {
@@ -14,5 +15,9 @@ export class PoolReports extends Entity {
 
   sharePrices(filter?: SharePricesReportFilter) {
     return new PoolSharePricesReport(this._root, this).report(filter)
+  }
+
+  shareYields(filter?: ShareYieldsReportFilter) {
+    return new PoolShareYieldsReport(this._root, this).report(filter)
   }
 }
