+++ src/index.ts
@@ -8,6 +8,7 @@ export * from './entities/OnOffRampManager.js'
 export * from './entities/Pool.js'
 export * from './entities/PoolNetwork.js'
 export * from './entities/Reports/PoolReports.js'
+export * from './entities/Reports/PoolShareYieldsReport.js'
 export * from './entities/Reports/PoolSharePricesReport.js'
 export * from './entities/ShareClass.js'
 export * from './entities/Vault.js'
